requirements.txt
flask==3.0.0
numpy==1.26.2
python-dotenv==1.0.0
sentence-transformers==2.2.2
